// src/app/modules/course/course.interface.ts
import { Types } from 'mongoose'

export type TPrerequisiteCourse = {
  course: Types.ObjectId
  isDeleted: boolean
}

export type TCourse = {
  title: string
  prefix: string
  code: number
  credits: number
  preRequisiteCourses: [TPrerequisiteCourse]
  isDelete?: boolean
}

export type TCourseFaculties = {
  course: Types.ObjectId
  faculties: [Types.ObjectId]
}